+++ src/BugsnagUnity/Native/Android/NativeInterface.cs
@@ -616,7 +616,8 @@ namespace BugsnagUnity
 
         public Dictionary<string, object> GetMetadata()
         {
-            return GetJavaMapData("getMetadata");
+            var metadata = GetJavaMapData("getMetadata");
+            return metadata;
         }
 
         public Dictionary<string, object> GetUser()
@@ -642,14 +643,134 @@ namespace BugsnagUnity
             CallNativeVoidMethod("clearMetadata", "(Ljava/lang/String;Ljava/lang/String;)V", new object[] { tab, key });
         }
 
-        public void AddMetadata(string tab, string key, string value)
+        public void AddMetadata(string section, IDictionary<string,object> metadata)
         {
-            if (tab == null || key == null)
+            if (section == null || metadata == null)
             {
                 return;
             }
-            CallNativeVoidMethod("addMetadata", "(Ljava/lang/String;Ljava/lang/String;Ljava/lang/Object;)V",
-                new object[] { tab, key, value });
+            CallNativeVoidMethod("addMetadata", "(Ljava/lang/String;Ljava/util/Map;)V",
+            new object[] { section, DictionaryToJavaMap(metadata) });           
+        }
+
+        internal static AndroidJavaObject DictionaryToJavaMap(IDictionary<string, object> inputDict)
+        {
+            AndroidJavaObject map = new AndroidJavaObject("java.util.HashMap");
+            if (inputDict != null)
+            {
+                foreach (var entry in inputDict)
+                {
+                    var key = new AndroidJavaObject("java.lang.String", entry.Key);
+
+                    if (entry.Value == null)
+                    {
+                        map.Call<AndroidJavaObject>("put", key, null);
+                    }
+                    else if (entry.Value is IDictionary)
+                    {
+                        if (entry.Value is IDictionary<string, object> dictionary)
+                        {
+                            map.Call<AndroidJavaObject>("put", key, DictionaryToJavaMap(dictionary));
+                        }
+                        else
+                        {
+                            var convertedDictionary = ConvertIfPoss(entry.Value);
+                            if (convertedDictionary != null)
+                            {
+                                map.Call<AndroidJavaObject>("put", key, DictionaryToJavaMap(convertedDictionary));
+                            }
+                        }
+                    }
+                    else if (IsListOrArray(entry.Value))
+                    {
+                        map.Call<AndroidJavaObject>("put", key, GetJavaArrayListFromCollection(entry.Value));
+                    }
+                    else
+                    {
+                        map.Call<AndroidJavaObject>("put", key, GetAndroidObjectFromBasicObject(entry.Value));
+                    }
+                }
+            }
+            return map;
+        }
+
+        private static IDictionary<string, object> ConvertIfPoss(object o)
+        {
+            var stringDict = o as IDictionary<string, object>;
+            if (stringDict != null)
+            {
+                return stringDict;
+            }
+
+            var dict = o as IDictionary;
+
+            if (dict != null)
+            {
+                stringDict = new Dictionary<string, object>();
+                foreach (var key in dict.Keys)
+                {
+                    stringDict[key?.ToString() ?? ""] = dict[key];
+                }
+                return stringDict;
+            }
+            return null;
+        }
+
+        private static bool IsListOrArray(object theObject)
+        {
+            var oType = theObject.GetType();
+            return (oType.IsGenericType && oType.GetGenericTypeDefinition() == typeof(List<>)) || oType.IsArray;
+        }
+
+        private static AndroidJavaObject GetJavaArrayListFromCollection(object theObject)
+        {
+            var collection = (IEnumerable)theObject;
+            var arrayList = new AndroidJavaObject("java.util.ArrayList");
+            foreach (var item in collection)
+            {
+                arrayList.Call<Boolean>("add", GetAndroidObjectFromBasicObject(item));
+            }
+            return arrayList;
+        }
+
+        private static AndroidJavaObject GetAndroidObjectFromBasicObject(object theObject)
+        {
+            if (theObject == null)
+            {
+                return null;
+            }
+            string nativeClass;
+            if (theObject.GetType() == typeof(string))
+            {
+                nativeClass = "java.lang.String";
+            }
+            else if (theObject.GetType() == typeof(bool))
+            {
+                nativeClass = "java.lang.Boolean";
+            }
+            else if (theObject.GetType() == typeof(int))
+            {
+                nativeClass = "java.lang.Integer";
+            }
+            else if (theObject.GetType() == typeof(double))
+            {
+                nativeClass = "java.lang.Double";
+            }
+            else if (theObject.GetType() == typeof(float))
+            {
+                nativeClass = "java.lang.Float";
+            }
+            else if (theObject.GetType() == typeof(long))
+            {
+                nativeClass = "java.lang.Long";
+            }
+            else
+            {
+                return new AndroidJavaObject("java.lang.String", theObject.ToString());
+            }
+
+            return new AndroidJavaObject(nativeClass, theObject);
+            
         }
 
         public void LeaveBreadcrumb(string name, string type, IDictionary<string, object> metadata)
@@ -665,7 +786,7 @@ namespace BugsnagUnity
             }
             if (PushLocalFrame())
             {
-                using (AndroidJavaObject map = BuildJavaMapDisposable(metadata))
+                using (AndroidJavaObject map = DictionaryToJavaMap(metadata))
                 {
                     CallNativeVoidMethod("leaveBreadcrumb", "(Ljava/lang/String;Ljava/lang/String;Ljava/util/Map;)V",
                         new object[] { name, type, map });
@@ -931,23 +1052,6 @@ namespace BugsnagUnity
 
             string timestamp = AndroidJNI.CallStringMethod(javaBreadcrumb, BreadcrumbGetTimestamp, new jvalue[] { });
             return new Breadcrumb(message, timestamp, typeName, metadata);
-        }
-
-        internal static AndroidJavaObject BuildJavaMapDisposable(IDictionary<string, object> src)
-        {
-            AndroidJavaObject map = new AndroidJavaObject("java.util.HashMap");
-            if (src != null)
-            {
-                foreach (var entry in src)
-                {
-                    using (AndroidJavaObject key = new AndroidJavaObject("java.lang.String", entry.Key))
-                    using (AndroidJavaObject value = new AndroidJavaObject("java.lang.String", entry.Value == null ? "null" : entry.Value.ToString()))
-                    {
-                        map.Call<AndroidJavaObject>("put", key, value);
-                    }
-                }
-            }
-            return map;
         }
 
         internal static bool IsUnity2019OrNewer()