+++ src/BugsnagUnity/SessionTracker.cs
@@ -101,8 +101,10 @@ namespace BugsnagUnity
             if (Client.Configuration.Endpoints.IsValid)
             {
                 var payload = new SessionReport(Client.Configuration, session);
-                Client.PayloadManager.AddPendingPayload(payload);               
-                Client.Send(payload);
+                if (Client.PayloadManager.AddPendingPayload(payload))
+                {
+                    Client.Send(payload);
+                }
             }
             else
             {